refactor(solver): extract shallow-subtype engine from intern/normalize.rs
Move the self-contained shallow subtype family (is_subtype_shallow and its
template/object/function helpers) out of intern/normalize.rs into a new
intern/shallow_subtype.rs. normalize.rs keeps the canonical reduction passes
(reduce_union_subtypes / reduce_intersection_subtypes) that call into the
engine. No behavior change: pure method relocation with pub(super) visibility
for the two cross-module entry points.

normalize.rs 2010 -> 1431 lines (under the 2000-line cap).

Refs #13417
